refact: remove unused file, clean files
